refactor(control-panel): colocate files and add barrel export
